requirements.txt
langchain-community
pypdf
langchain-huggingface
faiss-cpu
langchain_mongodb
Flask
python-dotenv
beautifulsoup4
python-pptx